move main to app folder
